Extract match handlers to separate module
